Moved Player to the table module since it is now a part of Table and deleted the Player module.
