+++ crapssim/__init__.py
@@ -1,7 +1,6 @@
-__all__ = ["table", "player", "dice", "strategy", "bet"]
+__all__ = ["table", "dice", "strategy", "bet"]
 
 from crapssim.dice import Dice
-from crapssim.player import Player
-from crapssim.table import Table
+from crapssim.table import Table, Player
 from . import bet
 from . import strategy
